Fetch ERA5Land time-series for a list of coordinates
This is a wrapper for the ECWMF's climate data store API. It reads in a list of coordinates from a csv file, then for each, makes API calls to get time series csv files for a list of variables. I am using this as an an excuse to learn python. There isn't much for error checking!